refactor( 🔧 ): reorganize directories
