+++ src/classes/cards/crops/index.ts
@@ -1,0 +1,4 @@
+import ChilliCard from "./chilli/ChilliCard";
+import CornCard from "./corn/CornCard";
+
+export { CornCard, ChilliCard };
