+++ src/classes/cards/goods/index.ts
@@ -1,0 +1,4 @@
+import CactusCard from "./cactus/CactusCard";
+import MagueyCard from "./maguey/MagueyCard";
+
+export { CactusCard, MagueyCard };
