+++ src/classes/decks/index.ts
@@ -1,0 +1,4 @@
+import CropDeckCreator from "./CropDeckCreator";
+import GoodDeckCreator from "./GoodDeckCreator";
+
+export { CropDeckCreator, GoodDeckCreator };
