+++ src/classes/interactors/crops/chilli/ChilliEdgeSlotInteractor.ts
@@ -1,15 +1,14 @@
-import SlotInteractor from "../../../slots/SlotInteractor";
 import {
   MAX_CARDS_PER_EDGE_SLOT,
   MAX_CARDS_PER_MILPA_SLOT,
   SlotType,
 } from "../../../../common";
-import Slot from "../../../slots/Slot";
-import Card from "../../Card";
-import CactusCard from "./CactusCard";
-import EdgeSlot from "../../../slots/EdgeSlot";
+import Card from "../../../cards/Card";
+import { SlotInteractor } from "../../../../interfaces";
+import { ChilliCard } from "../../../cards/crops";
+import { EdgeSlot, Slot } from "../../../slots";
 
-class CactusEdgeSlotInteractor implements SlotInteractor {
+class ChilliEdgeSlotInteractor implements SlotInteractor {
   private maxCardsPerMilpaSlotModifier = MAX_CARDS_PER_MILPA_SLOT;
 
   private maxCardsPerEdgeSlotModifier = MAX_CARDS_PER_EDGE_SLOT;
@@ -22,12 +21,10 @@ class CactusEdgeSlotInteractor implements SlotInteractor {
 
   pushToSlot(slot: Slot): Slot {
     const newCards: Card[] = [];
-    // TODO Implement Card Creator Singleton
     if (slot.type === SlotType.MILPA && this.canInteractWithSlot(slot))
-      newCards.push(new CactusCard());
-    // TODO Create right slot
+      newCards.push(new ChilliCard());
     return new EdgeSlot(newCards);
   }
 }
 
-export default CactusEdgeSlotInteractor;
+export default ChilliEdgeSlotInteractor;
