+++ src/classes/interactors/crops/chilli/ChilliMilpaSlotInteractor.ts
@@ -1,15 +1,14 @@
-import SlotInteractor from "../../../slots/SlotInteractor";
 import {
   MAX_CARDS_PER_EDGE_SLOT,
   MAX_CARDS_PER_MILPA_SLOT,
   SlotType,
 } from "../../../../common";
-import Slot from "../../../slots/Slot";
-import Card from "../../Card";
-import MagueyCard from "./MagueyCard";
-import EdgeSlot from "../../../slots/EdgeSlot";
+import Card from "../../../cards/Card";
+import { SlotInteractor } from "../../../../interfaces";
+import { ChilliCard } from "../../../cards/crops";
+import { MilpaSlot, Slot } from "../../../slots";
 
-class MagueyEdgeSlotInteractor implements SlotInteractor {
+class ChilliMilpaSlotInteractor implements SlotInteractor {
   private maxCardsPerMilpaSlotModifier = MAX_CARDS_PER_MILPA_SLOT;
 
   private maxCardsPerEdgeSlotModifier = MAX_CARDS_PER_EDGE_SLOT;
@@ -22,12 +21,10 @@ class MagueyEdgeSlotInteractor implements SlotInteractor {
 
   pushToSlot(slot: Slot): Slot {
     const newCards: Card[] = [];
-    // TODO Implement Card Creator Singleton
     if (slot.type === SlotType.MILPA && this.canInteractWithSlot(slot))
-      newCards.push(new MagueyCard());
-    // TODO Create right slot
-    return new EdgeSlot(newCards);
+      newCards.push(new ChilliCard());
+    return new MilpaSlot(newCards);
   }
 }
 
-export default MagueyEdgeSlotInteractor;
+export default ChilliMilpaSlotInteractor;
