+++ src/classes/interactors/crops/corn/CornEdgeSlotInteractor.ts
@@ -1,15 +1,14 @@
-import SlotInteractor from "../../../slots/SlotInteractor";
 import {
   MAX_CARDS_PER_EDGE_SLOT,
   MAX_CARDS_PER_MILPA_SLOT,
   SlotType,
 } from "../../../../common";
-import Slot from "../../../slots/Slot";
-import Card from "../../Card";
-import MilpaSlot from "../../../slots/MilpaSlot";
-import CactusCard from "./CactusCard";
+import Card from "../../../cards/Card";
+import { SlotInteractor } from "../../../../interfaces";
+import { CornCard } from "../../../cards/crops";
+import { EdgeSlot, Slot } from "../../../slots";
 
-class CactusMilpaSlotInteractor implements SlotInteractor {
+class CornEdgeSlotInteractor implements SlotInteractor {
   private maxCardsPerMilpaSlotModifier = MAX_CARDS_PER_MILPA_SLOT;
 
   private maxCardsPerEdgeSlotModifier = MAX_CARDS_PER_EDGE_SLOT;
@@ -24,10 +23,10 @@ class CactusMilpaSlotInteractor implements SlotInteractor {
     const newCards: Card[] = [];
     // TODO Implement Card Creator Singleton
     if (slot.type === SlotType.MILPA && this.canInteractWithSlot(slot))
-      newCards.push(new CactusCard());
+      newCards.push(new CornCard());
     // TODO Create right slot
-    return new MilpaSlot(newCards);
+    return new EdgeSlot(newCards);
   }
 }
 
-export default CactusMilpaSlotInteractor;
+export default CornEdgeSlotInteractor;
