+++ src/classes/interactors/crops/corn/CornMilpaSlotInteractor.ts
@@ -1,15 +1,14 @@
-import SlotInteractor from "../../../slots/SlotInteractor";
 import {
   MAX_CARDS_PER_EDGE_SLOT,
   MAX_CARDS_PER_MILPA_SLOT,
   SlotType,
 } from "../../../../common";
-import Slot from "../../../slots/Slot";
-import Card from "../../Card";
-import CornCard from "./CornCard";
-import EdgeSlot from "../../../slots/EdgeSlot";
+import Card from "../../../cards/Card";
+import { SlotInteractor } from "../../../../interfaces";
+import { CornCard } from "../../../cards/crops";
+import { MilpaSlot, Slot } from "../../../slots";
 
-class CornEdgeSlotInteractor implements SlotInteractor {
+class CornMilpaSlotInteractor implements SlotInteractor {
   private maxCardsPerMilpaSlotModifier = MAX_CARDS_PER_MILPA_SLOT;
 
   private maxCardsPerEdgeSlotModifier = MAX_CARDS_PER_EDGE_SLOT;
@@ -26,8 +25,8 @@ class CornEdgeSlotInteractor implements SlotInteractor {
     if (slot.type === SlotType.MILPA && this.canInteractWithSlot(slot))
       newCards.push(new CornCard());
     // TODO Create right slot
-    return new EdgeSlot(newCards);
+    return new MilpaSlot(newCards);
   }
 }
 
-export default CornEdgeSlotInteractor;
+export default CornMilpaSlotInteractor;
