+++ src/classes/interactors/crops/index.ts
@@ -1,0 +1,11 @@
+import ChilliEdgeSlotInteractor from "./chilli/ChilliEdgeSlotInteractor";
+import ChilliMilpaSlotInteractor from "./chilli/ChilliMilpaSlotInteractor";
+import CornEdgeSlotInteractor from "./corn/CornEdgeSlotInteractor";
+import CornMilpaSlotInteractor from "./corn/CornMilpaSlotInteractor";
+
+export {
+  CornMilpaSlotInteractor,
+  CornEdgeSlotInteractor,
+  ChilliMilpaSlotInteractor,
+  ChilliEdgeSlotInteractor,
+};
