+++ src/classes/interactors/goods/cactus/CactusMilpaSlotInteractor.ts
@@ -1,15 +1,14 @@
-import SlotInteractor from "../../../slots/SlotInteractor";
 import {
   MAX_CARDS_PER_EDGE_SLOT,
   MAX_CARDS_PER_MILPA_SLOT,
   SlotType,
 } from "../../../../common";
-import Slot from "../../../slots/Slot";
-import Card from "../../Card";
-import MilpaSlot from "../../../slots/MilpaSlot";
-import CornCard from "./CornCard";
+import Card from "../../../cards/Card";
+import { SlotInteractor } from "../../../../interfaces";
+import { CactusCard } from "../../../cards/goods";
+import { MilpaSlot, Slot } from "../../../slots";
 
-class CornMilpaSlotInteractor implements SlotInteractor {
+class CactusMilpaSlotInteractor implements SlotInteractor {
   private maxCardsPerMilpaSlotModifier = MAX_CARDS_PER_MILPA_SLOT;
 
   private maxCardsPerEdgeSlotModifier = MAX_CARDS_PER_EDGE_SLOT;
@@ -24,10 +23,10 @@ class CornMilpaSlotInteractor implements SlotInteractor {
     const newCards: Card[] = [];
     // TODO Implement Card Creator Singleton
     if (slot.type === SlotType.MILPA && this.canInteractWithSlot(slot))
-      newCards.push(new CornCard());
+      newCards.push(new CactusCard());
     // TODO Create right slot
     return new MilpaSlot(newCards);
   }
 }
 
-export default CornMilpaSlotInteractor;
+export default CactusMilpaSlotInteractor;
