+++ src/classes/interactors/goods/index.ts
@@ -1,0 +1,11 @@
+import CactusEdgeSlotInteractor from "./cactus/CactusEdgeSlotInteractor";
+import CactusMilpaSlotInteractor from "./cactus/CactusMilpaSlotInteractor";
+import MagueyMilpaSlotInteractor from "./maguey/MagueyMilpaSlotInteractor";
+import MagueyEdgeSlotInteractor from "./maguey/MagueyEdgeSlotInteractor";
+
+export {
+  CactusMilpaSlotInteractor,
+  CactusEdgeSlotInteractor,
+  MagueyEdgeSlotInteractor,
+  MagueyMilpaSlotInteractor,
+};
