+++ src/classes/score/crops/index.ts
@@ -1,0 +1,4 @@
+import CornScoreCalculator from "./corn/CornScoreCalculator";
+import ChilliScoreCalculator from "./chilli/ChilliScoreCalculator";
+
+export { CornScoreCalculator, ChilliScoreCalculator };
