+++ src/classes/score/goods/index.ts
@@ -1,0 +1,4 @@
+import CactusScoreCalculator from "./cactus/CactusScoreCalculator";
+import MagueyScoreCalculator from "./maguey/MagueyScoreCalculator";
+
+export { CactusScoreCalculator, MagueyScoreCalculator };
