+++ src/classes/slots/EdgeSlot.ts
@@ -2,7 +2,7 @@ import Slot from "./Slot";
 import { SlotType } from "../../common";
 
 class EdgeSlot extends Slot {
-  type: SlotType = SlotType.MILPA;
+  public readonly type: SlotType = SlotType.MILPA;
 }
 
 export default EdgeSlot;
