+++ src/classes/slots/MilpaSlot.ts
@@ -2,7 +2,7 @@ import Slot from "./Slot";
 import { SlotType } from "../../common";
 
 class MilpaSlot extends Slot {
-  type: SlotType = SlotType.MILPA;
+  public readonly type: SlotType = SlotType.MILPA;
 }
 
 export default MilpaSlot;
