+++ src/classes/slots/index.ts
@@ -1,0 +1,5 @@
+import EdgeSlot from "./EdgeSlot";
+import MilpaSlot from "./MilpaSlot";
+import Slot from "./Slot";
+
+export { EdgeSlot, MilpaSlot, Slot };
