+++ src/interfaces/index.ts
@@ -1,0 +1,4 @@
+import ScoreCalculator from "./ScoreCalculator";
+import SlotInteractor from "./SlotInteractor";
+
+export { SlotInteractor, ScoreCalculator };
